options page works now too 😃
