+++ lib/filterLocations.js
@@ -60,11 +60,7 @@ const doesIntersect = (list1, list2) => {
 function categoriesFilter(locationOptions) {
   const { location, options: { categories }, isValid } = locationOptions
   if (!isValid || categories.length === 0) return locationOptions
-  // console.log(location)
-  // console.log(location)
   const locationCategories = relevantTaxonomies(location.services)
-  // const locationHasCategory = category => locationCategories.includes(category)
-  // console.log(locationCategories)
 
   const isStillValid = doesIntersect(categories, locationCategories)
   return { ...locationOptions, isValid: isStillValid }
@@ -103,12 +99,10 @@ function genderFilter(locationOptions) {
 // LocationFilterOptions -> LocationFilterOptions
 function hoursFilter(locationOptions) {
   const { location, options, isValid } = locationOptions
-  console.log(location.name)
   if (!isValid || options.hours === 'all') return locationOptions
   const { services } = location
 
   const isStillValid = isOpen(services)
-  console.log(isStillValid)
 
   return {...locationOptions, isValid: isStillValid }
 }
@@ -121,7 +115,6 @@ function filterByOptions(locationAndFilterOptions) {
     genderFilter,
     hoursFilter
   )(locationAndFilterOptions)
-  // console.log(something)
   return something.isValid
 }
